requirements.txt
line-bot-sdk
flask
openai 
gunicorn
firebase-admin
